refactor: safe Tier 0+1 — backend partial-class splits + util tests
Zero behaviour change. C# partial classes compile to identical IL; backend
integration + unit tests verify no regression. New unit tests are additive
(test infrastructure for mobile is new; web Vitest existed already).

Backend cosmetic splits (compile-identical):
- AppDbContext.cs        655 → 104 LOC + 7 partials (Catalog, User, Reading,
                                                     UserBooks, Vocabulary, Ops,
                                                     Seo, Collections)
- VocabularyEndpoints.cs 1559 → 875 LOC + 6 partials (Stats, Settings, Pending,
                                                      Lookups, Clusters, Admin)
- AdminService.cs        977 → 131 LOC + 4 partials (Upload, Editions,
                                                     Chapters, UserUploads)

Each partial < 400 LOC, single domain per file. Largest non-migration backend
file dropped from 1559 → 875 LOC.

Tests added (Tier 0 — additive):
- Web: analytics.test.ts (15), dataEvents.test.ts (12), errorUtils.test.ts
  (5), formatTime.test.ts (8) — +40 cases
- Mobile (NEW Vitest infra): searchUtils.test.ts (18), features.test.ts (9),
  vocabStatsCache.test.ts (11) — +38 cases, vitest.config.ts +
  __mocks__/async-storage.ts (in-process AsyncStorage mock)

Verification:
- dotnet build: 0 warnings, 0 errors
- dotnet test UnitTests: 216/216 pass
- dotnet test Extraction.Tests: 313/313 pass
- dotnet test Search.Tests: 20/20 pass
- pnpm tsc (web): exit 0
- pnpm test (web): 474/474 pass
- pnpm tsc (mobile): exit 0
- pnpm test (mobile): 38/38 pass
