+++ backend/src/Api/Endpoints/VocabularyEndpoints.cs
@@ -13,7 +13,24 @@ using TextStack.Vocabulary.Contracts;
 
 namespace Api.Endpoints;
 
-public static class VocabularyEndpoints
+/// <summary>
+/// Vocabulary HTTP surface. Routes registered via <see cref="MapVocabularyEndpoints"/>;
+/// handlers are split across partial files by sub-domain to keep each
+/// file under ~300 LOC and reviewable in isolation:
+///
+///   - VocabularyEndpoints.Stats.cs     GetStats, GetDailyStats, ParseTzOffset
+///   - VocabularyEndpoints.Settings.cs  GetSettings, UpdateSettings, UnretireWord
+///   - VocabularyEndpoints.Pending.cs   GetPending, PromotePending, DismissPending
+///   - VocabularyEndpoints.Lookups.cs   GetLookups, PromoteLookup, DismissLookup
+///   - VocabularyEndpoints.Clusters.cs  GetClusters, StartClusterBonus, DismissCluster, CompleteCluster
+///   - VocabularyEndpoints.Admin.cs     BackfillDefinitions
+///
+/// Everything else (SaveWord + Words CRUD + Review + helpers + DTOs)
+/// stays in this file. Splits use C# `partial` — compile-identical to
+/// the original monolithic file. Shared helpers (TryGetAuth, ToDto,
+/// UpsertLookupAsync, FindUserWordAsync, QueueEnrichment) live here.
+/// </summary>
+public static partial class VocabularyEndpoints
 {
     private const int MaxWordsPerUser = 5000;
     private const int MaxDistractorPoolSize = 200;
@@ -639,263 +656,6 @@ public static class VocabularyEndpoints
             newInterval, word.NextReviewAt, word.TotalReviews, word.CorrectReviews));
     }
 
-    // --- Stats ---
-
-    private static async Task<IResult> GetStats(
-        HttpContext httpContext,
-        AuthService authService,
-        IAppDbContext db,
-        WeeklyBudgetService weeklyBudget,
-        DailyCapService dailyCap,
-        CancellationToken ct)
-    {
-        if (!TryGetAuth(httpContext, authService, out var userId, out var siteId))
-            return Results.Unauthorized();
-
-        var now = DateTimeOffset.UtcNow;
-        var todayStart = new DateTimeOffset(now.Date, TimeSpan.Zero);
-
-        // Base scope: everything the user owns. Stage breakdown + totalWords
-        // keep retired rows (user still wants to see "2000 mastered" even if
-        // they no longer appear in the queue).
-        var words = db.VocabularyWords
-            .Where(w => w.UserId == userId && w.SiteId == siteId);
-
-        var totalWords = await words.CountAsync(ct);
-
-        var byStage = await words
-            .GroupBy(w => w.Stage)
-            .Select(g => new { Stage = g.Key, Count = g.Count() })
-            .ToListAsync(ct);
-
-        var stageDict = byStage.ToDictionary(s => s.Stage, s => s.Count);
-        // dueNow intentionally excludes retired rows — matches what the review
-        // queue actually returns, so the banner never over-promises work.
-        var dueNow = await words.CountAsync(w => !w.IsRetired && w.NextReviewAt <= now, ct);
-        var retiredCount = await words.CountAsync(w => w.IsRetired, ct);
-
-        // Single query for today's review stats
-        var todayStats = await db.VocabularyReviews
-            .Where(r => r.UserId == userId && r.SiteId == siteId && r.CreatedAt >= todayStart)
-            .GroupBy(_ => 1)
-            .Select(g => new
-            {
-                Total = g.Count(),
-                Correct = g.Count(r => r.IsCorrect),
-                Practice = g.Count(r => r.ReviewMode.StartsWith("practice_")),
-                PracticeCorrect = g.Count(r => r.ReviewMode.StartsWith("practice_") && r.IsCorrect),
-            })
-            .FirstOrDefaultAsync(ct);
-
-        var reviewedToday = todayStats?.Total ?? 0;
-        var correctToday = todayStats?.Correct ?? 0;
-        var practiceToday = todayStats?.Practice ?? 0;
-        var practiceCorrectToday = todayStats?.PracticeCorrect ?? 0;
-        var srsReviewedToday = reviewedToday - practiceToday;
-        var srsCorrectToday = correctToday - practiceCorrectToday;
-
-        // Single query for all-time review stats
-        var allStats = await db.VocabularyReviews
-            .Where(r => r.UserId == userId && r.SiteId == siteId)
-            .GroupBy(_ => 1)
-            .Select(g => new { Total = g.Count(), Correct = g.Count(r => r.IsCorrect) })
-            .FirstOrDefaultAsync(ct);
-
-        var totalReviews = allStats?.Total ?? 0;
-        var totalCorrect = allStats?.Correct ?? 0;
-
-        // Streak: consecutive days with reviews (HashSet for O(1) lookup)
-        var reviewDays = (await db.VocabularyReviews
-            .Where(r => r.UserId == userId && r.SiteId == siteId)
-            .Select(r => r.CreatedAt.Date)
-            .Distinct()
-            .OrderByDescending(d => d)
-            .Take(365)
-            .ToListAsync(ct))
-            .ToHashSet();
-
-        var streak = 0;
-        var checkDate = now.Date;
-        if (!reviewDays.Contains(checkDate))
-            checkDate = checkDate.AddDays(-1);
-        while (reviewDays.Contains(checkDate))
-        {
-            streak++;
-            checkDate = checkDate.AddDays(-1);
-        }
-
-        var wordsByBook = await words
-            .Where(w => w.BookTitle != null)
-            .GroupBy(w => new { w.EditionId, w.UserBookId, w.BookTitle })
-            .Select(g => new { g.Key.EditionId, g.Key.UserBookId, g.Key.BookTitle, Count = g.Count() })
-            .OrderByDescending(b => b.Count)
-            .Take(20)
-            .ToListAsync(ct);
-
-        var weeklyProgress = ToDto(await weeklyBudget.GetProgressAsync(userId, siteId, ct));
-        var capStatus = await dailyCap.GetStatusAsync(userId, siteId, ct);
-        var pendingCount = await db.PendingVocabularyWords
-            .CountAsync(p => p.UserId == userId && p.SiteId == siteId, ct);
-        var lookupCount = await db.WordLookups
-            .CountAsync(l => l.UserId == userId && l.SiteId == siteId, ct);
-        var clusterCount = await db.WordClusters
-            .CountAsync(c => c.UserId == userId && c.SiteId == siteId
-                          && !c.IsDismissed && c.CompletedAt == null, ct);
-
-        return Results.Ok(new
-        {
-            totalWords,
-            byStage = new
-            {
-                @new = stageDict.GetValueOrDefault(0),
-                recognition = stageDict.GetValueOrDefault(1),
-                recall = stageDict.GetValueOrDefault(2),
-                context = stageDict.GetValueOrDefault(3),
-                mastered = stageDict.GetValueOrDefault(4),
-            },
-            dueNow,
-            retiredCount,
-            pendingCount,
-            lookupCount,
-            clusterCount,
-            weeklyProgress,
-            dailyCap = new { used = capStatus.Used, cap = capStatus.Cap, remaining = capStatus.Remaining },
-            reviewedToday,
-            correctRateToday = reviewedToday > 0 ? Math.Round((double)correctToday / reviewedToday * 100, 1) : 0,
-            srsReviewedToday,
-            srsCorrectRateToday = srsReviewedToday > 0 ? Math.Round((double)srsCorrectToday / srsReviewedToday * 100, 1) : 0,
-            practicedToday = practiceToday,
-            practiceCorrectRateToday = practiceToday > 0 ? Math.Round((double)practiceCorrectToday / practiceToday * 100, 1) : 0,
-            totalReviews,
-            overallCorrectRate = totalReviews > 0 ? Math.Round((double)totalCorrect / totalReviews * 100, 1) : 0,
-            streak,
-            wordsByBook,
-        });
-    }
-
-    // --- Daily Stats ---
-
-    private static TimeSpan ParseTzOffset(string? tz)
-    {
-        if (string.IsNullOrEmpty(tz)) return TimeSpan.Zero;
-        if (int.TryParse(tz, out var minutes))
-            return TimeSpan.FromMinutes(minutes);
-        return TimeSpan.Zero;
-    }
-
-    private static async Task<IResult> GetDailyStats(
-        HttpContext httpContext,
-        AuthService authService,
-        IAppDbContext db,
-        [FromQuery] DateTimeOffset? from,
-        [FromQuery] DateTimeOffset? to,
-        [FromQuery] string? tz,
-        CancellationToken ct)
-    {
-        if (!TryGetAuth(httpContext, authService, out var userId, out var siteId))
-            return Results.Unauthorized();
-
-        var tzOffset = ParseTzOffset(tz);
-        var now = DateTimeOffset.UtcNow;
-        var start = from ?? now.AddDays(-365);
-        var end = to ?? now;
-
-        // Reviews per day
-        var reviews = await db.VocabularyReviews
-            .Where(r => r.UserId == userId && r.SiteId == siteId
-                && r.CreatedAt >= start && r.CreatedAt <= end)
-            .Select(r => new { r.CreatedAt, r.IsCorrect, r.ReviewMode })
-            .ToListAsync(ct);
-
-        var reviewsByDay = reviews
-            .GroupBy(r => r.CreatedAt.ToOffset(tzOffset).Date)
-            .ToDictionary(
-                g => g.Key,
-                g => new
-                {
-                    ReviewCount = g.Count(),
-                    CorrectCount = g.Count(r => r.IsCorrect),
-                    PracticeCount = g.Count(r => r.ReviewMode.StartsWith("practice_")),
-                    SrsCount = g.Count(r => !r.ReviewMode.StartsWith("practice_")),
-                });
-
-        // Words added per day
-        var words = await db.VocabularyWords
-            .Where(w => w.UserId == userId && w.SiteId == siteId
-                && w.CreatedAt >= start && w.CreatedAt <= end)
-            .Select(w => w.CreatedAt)
-            .ToListAsync(ct);
-
-        var wordsByDay = words
-            .GroupBy(d => d.ToOffset(tzOffset).Date)
-            .ToDictionary(g => g.Key, g => g.Count());
-
-        // Merge all dates
-        var allDates = reviewsByDay.Keys.Union(wordsByDay.Keys).OrderBy(d => d);
-
-        var result = allDates.Select(date =>
-        {
-            reviewsByDay.TryGetValue(date, out var r);
-            return new
-            {
-                date,
-                wordsAdded = wordsByDay.GetValueOrDefault(date, 0),
-                reviewCount = r?.ReviewCount ?? 0,
-                correctCount = r?.CorrectCount ?? 0,
-                practiceCount = r?.PracticeCount ?? 0,
-                srsCount = r?.SrsCount ?? 0,
-            };
-        }).ToList();
-
-        return Results.Ok(result);
-    }
-
-    // --- Admin: Backfill Definitions ---
-
-    private static async Task<IResult> BackfillDefinitions(
-        IAppDbContext db,
-        IDefinitionEnricher enricher,
-        ILogger<IAppDbContext> logger,
-        CancellationToken ct)
-    {
-        var words = await db.VocabularyWords
-            .Where(w => w.Definition == null || w.Definition == "")
-            .OrderBy(w => w.CreatedAt)
-            .Take(500)
-            .ToListAsync(ct);
-
-        var enriched = 0;
-        var failed = 0;
-
-        foreach (var w in words)
-        {
-            try
-            {
-                var def = await enricher.FetchDefinitionAsync(w.Word, w.Language, ct);
-                if (def != null)
-                {
-                    w.Definition = def;
-                    enriched++;
-                }
-                else
-                {
-                    failed++;
-                }
-                // Rate limit: Free Dictionary API
-                await Task.Delay(200, ct);
-            }
-            catch (Exception ex)
-            {
-                logger.LogWarning(ex, "Backfill failed for word {Word}", w.Word);
-                failed++;
-            }
-        }
-
-        await db.SaveChangesAsync(ct);
-
-        return Results.Ok(new { total = words.Count, enriched, failed });
-    }
-
     // --- Reader Vocab (lightweight word+stage list) ---
 
     private static async Task<IResult> GetReaderVocab(
@@ -1018,450 +778,6 @@ public static class VocabularyEndpoints
     private static WeeklyProgressDto ToDto(WeeklyProgress p) =>
         new(p.Used, p.Budget, p.Remaining, p.ResetAt);
 
-    // --- Settings (Phase 1 anti-spiral) ---
-
-    private static async Task<IResult> GetSettings(
-        HttpContext httpContext,
-        AuthService authService,
-        IAppDbContext db,
-        CancellationToken ct)
-    {
-        if (!TryGetAuth(httpContext, authService, out var userId, out var siteId))
-            return Results.Unauthorized();
-
-        var settings = await db.UserVocabularySettings
-            .FirstOrDefaultAsync(s => s.UserId == userId && s.SiteId == siteId, ct);
-
-        // First-time read: return defaults without persisting — settings row is
-        // created lazily on first PUT to avoid a write on every new user.
-        return Results.Ok(new VocabSettingsDto(
-            DailyNewCap: settings?.DailyNewCap ?? 15,
-            WeeklyReviewBudget: settings?.WeeklyReviewBudget ?? 70,
-            FrequencyFilterEnabled: settings?.FrequencyFilterEnabled ?? true,
-            ClusteringEnabled: settings?.ClusteringEnabled ?? true,
-            AutoRetireEnabled: settings?.AutoRetireEnabled ?? true));
-    }
-
-    private static async Task<IResult> UpdateSettings(
-        [FromBody] VocabSettingsDto request,
-        HttpContext httpContext,
-        AuthService authService,
-        IAppDbContext db,
-        CancellationToken ct)
-    {
-        if (!TryGetAuth(httpContext, authService, out var userId, out var siteId))
-            return Results.Unauthorized();
-
-        if (request.DailyNewCap is < 5 or > 100)
-            return Results.BadRequest("DailyNewCap must be 5–100");
-        if (request.WeeklyReviewBudget is < 10 or > 500)
-            return Results.BadRequest("WeeklyReviewBudget must be 10–500");
-
-        var settings = await db.UserVocabularySettings
-            .FirstOrDefaultAsync(s => s.UserId == userId && s.SiteId == siteId, ct);
-        var now = DateTimeOffset.UtcNow;
-
-        if (settings is null)
-        {
-            settings = new UserVocabularySettings
-            {
-                UserId = userId,
-                SiteId = siteId,
-                CreatedAt = now,
-            };
-            db.UserVocabularySettings.Add(settings);
-        }
-
-        settings.DailyNewCap = request.DailyNewCap;
-        settings.WeeklyReviewBudget = request.WeeklyReviewBudget;
-        settings.FrequencyFilterEnabled = request.FrequencyFilterEnabled;
-        settings.ClusteringEnabled = request.ClusteringEnabled;
-        settings.AutoRetireEnabled = request.AutoRetireEnabled;
-        settings.UpdatedAt = now;
-
-        await db.SaveChangesAsync(ct);
-        return Results.Ok(request);
-    }
-
-    // --- Unretire (Phase 1) ---
-
-    private static async Task<IResult> UnretireWord(
-        Guid id,
-        HttpContext httpContext,
-        AuthService authService,
-        IAppDbContext db,
-        CancellationToken ct)
-    {
-        if (!TryGetAuth(httpContext, authService, out var userId, out var siteId))
-            return Results.Unauthorized();
-
-        var word = await FindUserWordAsync(db, id, userId, siteId, ct);
-        if (word == null) return Results.NotFound();
-
-        if (!word.IsRetired) return Results.Ok(ToDto(word));
-
-        // Drop back to Stage 3 (Context), not Stage 0 — the user once mastered
-        // this word, so we're only asking for one resurfacing review, not
-        // starting from scratch. ConsecutiveCorrect resets so the 3/14d
-        // retirement rule has to be re-earned.
-        var now = DateTimeOffset.UtcNow;
-        word.IsRetired = false;
-        word.RetiredAt = null;
-        word.RetiredReason = null;
-        word.Stage = 3;
-        word.ConsecutiveCorrect = 0;
-        word.IntervalDays = 1;
-        word.NextReviewAt = now;
-        word.UpdatedAt = now;
-
-        await db.SaveChangesAsync(ct);
-        return Results.Ok(ToDto(word));
-    }
-
-    // --- Pending Buffer (Phase 2) ---
-
-    private static async Task<IResult> GetPending(
-        HttpContext httpContext,
-        AuthService authService,
-        IAppDbContext db,
-        DailyCapService dailyCap,
-        CancellationToken ct)
-    {
-        if (!TryGetAuth(httpContext, authService, out var userId, out var siteId))
-            return Results.Unauthorized();
-
-        var items = await db.PendingVocabularyWords
-            .Where(p => p.UserId == userId && p.SiteId == siteId)
-            .OrderByDescending(p => p.Priority)
-            .ThenBy(p => p.CreatedAt)
-            .Select(p => new PendingVocabWordDto(
-                p.Id, p.Word, p.Language, p.Translation, p.Definition,
-                p.EditionId, p.ChapterId, p.UserBookId,
-                p.Sentence, p.BookTitle, p.Priority, p.Source, p.CreatedAt))
-            .ToListAsync(ct);
-
-        var cap = await dailyCap.GetStatusAsync(userId, siteId, ct);
-        return Results.Ok(new PendingListResponse(items, cap.Used, cap.Cap, cap.Remaining));
-    }
-
-    // Manual promote — bypasses the daily cap. User explicitly asked for this
-    // word to skip the queue; respect it even if today is "full".
-    private static async Task<IResult> PromotePending(
-        Guid id,
-        HttpContext httpContext,
-        AuthService authService,
-        IAppDbContext db,
-        DailyCapService dailyCap,
-        IServiceScopeFactory scopeFactory,
-        ILogger<IAppDbContext> logger,
-        CancellationToken ct)
-    {
-        if (!TryGetAuth(httpContext, authService, out var userId, out var siteId))
-            return Results.Unauthorized();
-
-        var pending = await db.PendingVocabularyWords
-            .FirstOrDefaultAsync(p => p.Id == id && p.UserId == userId && p.SiteId == siteId, ct);
-        if (pending == null) return Results.NotFound();
-
-        // Capture enrichment inputs before PromoteAsync disposes the pending row.
-        var wordText = pending.Word;
-        var lang = pending.Language;
-        var def = pending.Definition;
-        var sent = pending.Sentence;
-
-        var now = DateTimeOffset.UtcNow;
-        var promoted = await dailyCap.PromoteAsync(pending, now, ct);
-
-        var nativeLang = await db.Users
-            .Where(u => u.Id == userId)
-            .Select(u => u.NativeLanguage)
-            .FirstOrDefaultAsync(ct);
-        if (!string.IsNullOrWhiteSpace(nativeLang))
-        {
-            QueueEnrichment(scopeFactory, logger, promoted.Id, wordText, lang, def, sent, nativeLang);
-        }
-
-        return Results.Ok(ToDto(promoted));
-    }
-
-    private static async Task<IResult> DismissPending(
-        Guid id,
-        HttpContext httpContext,
-        AuthService authService,
-        IAppDbContext db,
-        CancellationToken ct)
-    {
-        if (!TryGetAuth(httpContext, authService, out var userId, out var siteId))
-            return Results.Unauthorized();
-
-        var pending = await db.PendingVocabularyWords
-            .FirstOrDefaultAsync(p => p.Id == id && p.UserId == userId && p.SiteId == siteId, ct);
-        if (pending == null) return Results.NotFound();
-
-        db.PendingVocabularyWords.Remove(pending);
-        await db.SaveChangesAsync(ct);
-        return Results.NoContent();
-    }
-
-    // --- Lookups (Phase 3 F1) ---
-
-    private static async Task<IResult> GetLookups(
-        HttpContext httpContext,
-        AuthService authService,
-        IAppDbContext db,
-        [FromQuery] int? limit,
-        [FromQuery] int? offset,
-        CancellationToken ct)
-    {
-        if (!TryGetAuth(httpContext, authService, out var userId, out var siteId))
-            return Results.Unauthorized();
-
-        var take = Math.Clamp(limit ?? 50, 1, 200);
-        var skip = Math.Max(0, offset ?? 0);
-
-        var baseQuery = db.WordLookups
-            .Where(l => l.UserId == userId && l.SiteId == siteId);
-
-        var total = await baseQuery.CountAsync(ct);
-        var items = await baseQuery
-            .OrderByDescending(l => l.LastTappedAt)
-            .Skip(skip)
-            .Take(take)
-            .Select(l => new WordLookupDto(
-                l.Id, l.Word, l.Language, l.ZipfRank, l.TapCount,
-                l.Sentence, l.BookTitle, l.EditionId, l.ChapterId, l.UserBookId,
-                l.LastTranslation, l.FirstTappedAt, l.LastTappedAt))
-            .ToListAsync(ct);
-
-        return Results.Ok(new WordLookupListResponse(items, total));
-    }
-
-    // "Add anyway" — user overrides the frequency filter. Creates a VocabularyWord
-    // directly and drops the Lookup. Bypasses the daily cap on purpose: the user
-    // explicitly asked for this word.
-    private static async Task<IResult> PromoteLookup(
-        Guid id,
-        HttpContext httpContext,
-        AuthService authService,
-        IAppDbContext db,
-        IServiceScopeFactory scopeFactory,
-        ILogger<IAppDbContext> logger,
-        CancellationToken ct)
-    {
-        if (!TryGetAuth(httpContext, authService, out var userId, out var siteId))
-            return Results.Unauthorized();
-
-        var lookup = await db.WordLookups
-            .FirstOrDefaultAsync(l => l.Id == id && l.UserId == userId && l.SiteId == siteId, ct);
-        if (lookup == null) return Results.NotFound();
-
-        // Guard against a race where the word was saved via another path between
-        // the list render and the promote click.
-        var already = await db.VocabularyWords
-            .FirstOrDefaultAsync(w => w.UserId == userId && w.SiteId == siteId
-                && w.Word == lookup.Word && w.Language == lookup.Language, ct);
-        if (already != null)
-        {
-            db.WordLookups.Remove(lookup);
-            await db.SaveChangesAsync(ct);
-            return Results.Ok(ToDto(already));
-        }
-
-        // Hard ceiling still applies — Add Anyway bypasses the daily cap, not
-        // the 5000-word vocabulary limit. Counts both active SRS + pending
-        // so a user can't exceed the cap via the lookup bypass.
-        var count = await db.VocabularyWords.CountAsync(
-            w => w.UserId == userId && w.SiteId == siteId, ct);
-        count += await db.PendingVocabularyWords.CountAsync(
-            p => p.UserId == userId && p.SiteId == siteId, ct);
-        if (count >= MaxWordsPerUser)
-            return Results.Problem("Vocabulary limit reached (5000 words)", statusCode: 429);
-
-        var now = DateTimeOffset.UtcNow;
-        var entry = new VocabularyWord
-        {
-            Id = Guid.NewGuid(),
-            UserId = userId,
-            SiteId = siteId,
-            Word = lookup.Word,
-            Language = lookup.Language,
-            Translation = lookup.LastTranslation,
-            EditionId = lookup.EditionId,
-            ChapterId = lookup.ChapterId,
-            UserBookId = lookup.UserBookId,
-            Sentence = lookup.Sentence,
-            BookTitle = lookup.BookTitle,
-            ZipfRank = lookup.ZipfRank,
-            Source = "manual_add_anyway",
-            ActivatedAt = now,
-            Stage = 0,
-            IntervalDays = 0,
-            ConsecutiveCorrect = 0,
-            NextReviewAt = now,
-            TotalReviews = 0,
-            CorrectReviews = 0,
-            CreatedAt = now,
-            UpdatedAt = now,
-        };
-        db.VocabularyWords.Add(entry);
-        db.WordLookups.Remove(lookup);
-        await db.SaveChangesAsync(ct);
-
-        var nativeLang = await db.Users
-            .Where(u => u.Id == userId)
-            .Select(u => u.NativeLanguage)
-            .FirstOrDefaultAsync(ct);
-        if (!string.IsNullOrWhiteSpace(nativeLang))
-        {
-            QueueEnrichment(scopeFactory, logger, entry.Id, entry.Word, entry.Language,
-                entry.Definition, entry.Sentence, nativeLang);
-        }
-
-        return Results.Ok(ToDto(entry));
-    }
-
-    private static async Task<IResult> DismissLookup(
-        Guid id,
-        HttpContext httpContext,
-        AuthService authService,
-        IAppDbContext db,
-        CancellationToken ct)
-    {
-        if (!TryGetAuth(httpContext, authService, out var userId, out var siteId))
-            return Results.Unauthorized();
-
-        var lookup = await db.WordLookups
-            .FirstOrDefaultAsync(l => l.Id == id && l.UserId == userId && l.SiteId == siteId, ct);
-        if (lookup == null) return Results.NotFound();
-
-        db.WordLookups.Remove(lookup);
-        await db.SaveChangesAsync(ct);
-        return Results.NoContent();
-    }
-
-    // --- Clusters (F3 bonus round) ---
-
-    private static async Task<IResult> GetClusters(
-        HttpContext httpContext,
-        AuthService authService,
-        IAppDbContext db,
-        CancellationToken ct)
-    {
-        if (!TryGetAuth(httpContext, authService, out var userId, out var siteId))
-            return Results.Unauthorized();
-
-        var clusters = await db.WordClusters
-            .Where(c => c.UserId == userId
-                     && c.SiteId == siteId
-                     && !c.IsDismissed
-                     && c.CompletedAt == null)
-            .OrderByDescending(c => c.CreatedAt)
-            .Take(10)
-            .Select(c => new WordClusterDto(
-                c.Id, c.Title, c.Theme,
-                c.EditionId, c.UserBookId, c.BookTitle,
-                c.MemberCount, c.CohesionScore,
-                c.IsConfirmed, c.CreatedAt))
-            .ToListAsync(ct);
-
-        return Results.Ok(new { items = clusters });
-    }
-
-    private static async Task<IResult> StartClusterBonus(
-        Guid id,
-        HttpContext httpContext,
-        AuthService authService,
-        IAppDbContext db,
-        IReviewCardBuilder cardBuilder,
-        CancellationToken ct)
-    {
-        if (!TryGetAuth(httpContext, authService, out var userId, out var siteId))
-            return Results.Unauthorized();
-
-        var cluster = await db.WordClusters
-            .FirstOrDefaultAsync(c => c.Id == id && c.UserId == userId && c.SiteId == siteId, ct);
-        if (cluster == null) return Results.NotFound();
-        if (cluster.IsDismissed || cluster.CompletedAt != null) return Results.Conflict();
-
-        // Load cluster members. Retired members are skipped — they graduated.
-        var members = await db.VocabularyWords
-            .Where(w => w.ClusterId == id
-                     && w.UserId == userId
-                     && w.SiteId == siteId
-                     && !w.IsRetired)
-            .ToListAsync(ct);
-
-        if (members.Count == 0) return Results.NotFound();
-
-        // Distractor pool from user's other words (not in cluster).
-        var languages = members.Select(w => w.Language).Distinct().ToList();
-        var memberIds = members.Select(w => w.Id).ToHashSet();
-        var pool = await db.VocabularyWords
-            .Where(w => w.UserId == userId && w.SiteId == siteId
-                     && !memberIds.Contains(w.Id)
-                     && languages.Contains(w.Language))
-            .OrderBy(_ => EF.Functions.Random())
-            .Take(MaxDistractorPoolSize)
-            .Select(w => new DistractorPoolEntry(w.Word, w.Language))
-            .ToListAsync(ct);
-
-        var wordsForReview = members.Select(w => new WordForReview(
-            w.Id, w.Word, w.Language,
-            w.Translation, w.Definition,
-            w.Sentence, w.BookTitle, w.Hint,
-            w.Explanation, w.Distractors,
-            w.Stage, w.TotalReviews)).ToList();
-
-        var cards = cardBuilder.BuildCards(wordsForReview, pool);
-        var cardDtos = cards.Select(c => new ReviewCardDto(
-            c.WordId, c.Word, c.Translation, c.Definition,
-            c.ReviewMode, c.BlankSentence, c.OriginalSentence, c.BookTitle,
-            c.Hint, c.Explanation, c.IsNew, c.Options, c.CorrectOptionIndex)).ToList();
-
-        cluster.IsConfirmed = true;
-        await db.SaveChangesAsync(ct);
-
-        return Results.Ok(new { clusterId = cluster.Id, title = cluster.Title, cards = cardDtos });
-    }
-
-    private static async Task<IResult> DismissCluster(
-        Guid id,
-        HttpContext httpContext,
-        AuthService authService,
-        IAppDbContext db,
-        CancellationToken ct)
-    {
-        if (!TryGetAuth(httpContext, authService, out var userId, out var siteId))
-            return Results.Unauthorized();
-
-        var cluster = await db.WordClusters
-            .FirstOrDefaultAsync(c => c.Id == id && c.UserId == userId && c.SiteId == siteId, ct);
-        if (cluster == null) return Results.NotFound();
-
-        cluster.IsDismissed = true;
-        cluster.DismissedAt = DateTimeOffset.UtcNow;
-        await db.SaveChangesAsync(ct);
-        return Results.NoContent();
-    }
-
-    private static async Task<IResult> CompleteCluster(
-        Guid id,
-        HttpContext httpContext,
-        AuthService authService,
-        IAppDbContext db,
-        CancellationToken ct)
-    {
-        if (!TryGetAuth(httpContext, authService, out var userId, out var siteId))
-            return Results.Unauthorized();
-
-        var cluster = await db.WordClusters
-            .FirstOrDefaultAsync(c => c.Id == id && c.UserId == userId && c.SiteId == siteId, ct);
-        if (cluster == null) return Results.NotFound();
-
-        cluster.CompletedAt = DateTimeOffset.UtcNow;
-        await db.SaveChangesAsync(ct);
-        return Results.NoContent();
-    }
 }
 
 // --- DTOs ---
